restructure and rename components
