+++ src/app/[locale]/(main)/users/page.tsx
@@ -1,5 +1,5 @@
 import { unstable_setRequestLocale } from "next-intl/server";
-import { UserManagementScreen } from "@/screens/user-management";
+import { UsersScreen } from "@/screens/users";
 
 import type { BaseParams } from "@/types/params";
 
@@ -9,5 +9,5 @@ export default function HomePage({ params: { locale } }: HomePageProps) {
   // Initial value
   unstable_setRequestLocale(locale);
 
-  return <UserManagementScreen />;
+  return <UsersScreen />;
 }
