+++ src/locales/en/screens/home/home.json
