--- src/locales/en/screens/user-management-add.json
@@ -1,11 +1,0 @@
-{
-  "userManagementAdd": {
-    "title": "Add User Information",
-    "filters": {
-      "search": "Search"
-    },
-    "breadcrumbs": {
-      "usermanagement": "User Management System"
-    }
-  }
-}
