--- src/locales/en/screens/user-management.json
@@ -1,10 +1,0 @@
-{
-  "userManagement": {
-    "title": "User management",
-    "addUser": "Add User",
-    "filters": {
-      "search": "search",
-      "status": "status"
-    }
-  }
-}
