+++ src/locales/en/screens/users/users.json
@@ -1,0 +1,19 @@
+{
+  "users": {
+    "title": "User management",
+    "addUser": "Add User",
+    "filters": {
+      "search": "search",
+      "status": "status"
+    },
+    "new": {
+      "title": "Add User Information",
+      "filters": {
+        "search": "Search"
+      },
+      "breadcrumbs": {
+        "usermanagement": "User Management System"
+      }
+    }
+  }
+}
