+++ src/locales/index.ts
@@ -6,10 +6,8 @@ export default async function getLocales(locale: Locale = "en") {
     messages: {
       ...(await import(`./${locale}/common.json`)).default,
       screens: {
-        ...(await import(`./${locale}/screens/home.json`)).default,
-        ...(await import(`./${locale}/screens/user-management.json`)).default,
-        ...(await import(`./${locale}/screens/user-management-add.json`))
-          .default,
+        ...(await import(`./${locale}/screens/home/home.json`)).default,
+        ...(await import(`./${locale}/screens/users/users.json`)).default,
       },
     },
   };
