+++ src/locales/th/screens/home/home.json
