--- src/locales/th/screens/user-management-add.json
@@ -1,11 +1,0 @@
-{
-  "userManagementAdd": {
-    "title": "เพิ่มข้อมูลผู้ใช้งาน",
-    "filters": {
-      "search": "ค้นหา"
-    },
-    "breadcrumbs": {
-      "usermanagement": "ระบบจัดการผู้ใช้"
-    }
-  }
-}
