--- src/locales/th/screens/user-management.json
@@ -1,10 +1,0 @@
-{
-  "userManagement": {
-    "title": "ระบบจัดการผู้ใช้",
-    "addUser": "เพิ่มผู้ใช้งาน",
-    "filters": {
-      "search": "ค้นหา",
-      "status": "สถานะ"
-    }
-  }
-}
