+++ src/locales/th/screens/users/users.json
@@ -1,0 +1,19 @@
+{
+  "users": {
+    "title": "ระบบจัดการผู้ใช้",
+    "addUser": "เพิ่มผู้ใช้งาน",
+    "filters": {
+      "search": "ค้นหา",
+      "status": "สถานะ"
+    },
+    "new": {
+      "title": "เพิ่มข้อมูลผู้ใช้งาน",
+      "filters": {
+        "search": "ค้นหา"
+      },
+      "breadcrumbs": {
+        "usermanagement": "ระบบจัดการผู้ใช้"
+      }
+    }
+  }
+}
