--- src/screens/user-management/index.ts
@@ -1,1 +1,0 @@
-export { default as UserManagementScreen } from "./UserManagementScreen";
