+++ src/screens/users/index.ts
@@ -1,0 +1,2 @@
+export { default as UsersScreen } from "./UsersScreen";
+export type { FilterProps, QueryParams, QueryParamsAction } from "./types";
