+++ src/screens/users/new/NewUsersScreen.tsx
@@ -1,6 +1,5 @@
 "use client";
 
-import AddRoundedIcon from "@mui/icons-material/AddRounded";
 import Box from "@mui/material/Box";
 import Stack from "@mui/material/Stack";
 
@@ -8,12 +7,10 @@ import { useTranslations } from "next-intl";
 import { debounce } from "radash";
 import { useImmerReducer } from "use-immer";
 
-import Button from "@/components/button";
 import Header from "@/components/header";
 import Pagination from "@/components/pagination";
 import { ROUTE } from "@/constants/routes";
 import TableData from "@/components/table-data";
-import { Link } from "@/lib/navigation";
 
 import Filters from "./Filters";
 import { MOCK_DATA, MOCK_PAGINATION_DATA_FROM_API } from "./fixtures";
@@ -23,9 +20,9 @@ import { styles } from "./styles";
 import type { ChangeEvent, SyntheticEvent } from "react";
 import type { OrderType } from "@/types/sort";
 
-export default function UserManagementScreen() {
+export default function NewUsersScreen() {
   // Hooks
-  const t = useTranslations("screens.userManagement");
+  const t = useTranslations("screens.users.new");
   const tCommon = useTranslations("common");
   const [state, dispatch] = useImmerReducer(reducer, initialState);
 
@@ -35,11 +32,21 @@ export default function UserManagementScreen() {
       dispatch({ type: "UPDATE_FILTER", payload: { key: "name", value } });
     })();
   }
+
   // Variables
+  const breadcrumbs = [
+    {
+      href: ROUTE.users.path,
+      name: t("breadcrumbs.usermanagement"),
+    },
+    {
+      name: t("title"),
+    },
+  ];
+
   const filterMessages = {
     search: t("filters.search"),
     show: tCommon("pagination.show"),
-    status: t("filters.status"),
     perPage: tCommon("pagination.perPage"),
   };
 
@@ -52,21 +59,10 @@ export default function UserManagementScreen() {
   };
 
   // Functions
-
   function onChangeInput(
     event: ChangeEvent<HTMLInputElement | HTMLTextAreaElement>,
   ) {
     debouncedFilter(event.target.value);
-  }
-
-  function onChangeSelectStatus(_: SyntheticEvent, options: any) {
-    dispatch({
-      type: "UPDATE_FILTER",
-      payload: {
-        key: "status",
-        value: options?.value,
-      },
-    });
   }
 
   function onChangeSelectPagination(_: SyntheticEvent, options: any) {
@@ -79,7 +75,6 @@ export default function UserManagementScreen() {
     });
   }
 
-  // Event handler
   function handleSortColumn(key: string, orderBy: OrderType) {
     dispatch({ type: "UPDATE_SORT", payload: { key, value: orderBy } });
   }
@@ -96,21 +91,14 @@ export default function UserManagementScreen() {
 
   return (
     <Stack sx={styles.root}>
-      <Header text={t("title")}>
-        <Button
-          variant="contained"
-          sx={styles.buttonCreate}
-          startIcon={<AddRoundedIcon />}
-          component={Link}
-          href={ROUTE.userManagementAdd.path}
-        >
-          {t("addUser")}
-        </Button>
-      </Header>
+      <Header
+        text={t("title")}
+        breadcrumbs={breadcrumbs}
+        hrefBack={ROUTE.users.path}
+      />
       <Box sx={styles.card}>
         <Filters
           onChangeInput={onChangeInput}
-          onChangeSelectStatus={onChangeSelectStatus}
           onChangeSelectPagination={onChangeSelectPagination}
           messages={filterMessages}
         />
